pyproject.toml
[project]
name = "fromScratchAPI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "alembic-postgresql-enum==1.2.0",
    "fastapi==0.111.0",
    "fastapi_sqlalchemy",
    "factory-boy==3.3.1",
    "google-auth==2.29.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.1",
    "pytest==8.2.0",
    "pytest-asyncio==0.21.0",
    "pytest-bdd==7.3.0",
    "pytz==2022.7",
    "requests==2.31.0",
    "SQLAlchemy==2.0.36",
    "SQLAlchemy-Continuum==1.4.2",
    "uvicorn==0.29.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.1",
    "faker>=37.11.0",
    "httpx>=0.28.1",
    "pytest>=8.2.0",
    "pytest-bdd>=7.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.11.9",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
